+++ src/DataTransferObjectError.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace Spatie\DataObject;
+namespace Spatie\DataTransferObject;
 
 use TypeError;
 
-class DataObjectError extends TypeError
+class DataTransferObjectError extends TypeError
 {
-    public static function unknownProperties(array $properties, string $className): DataObjectError
+    public static function unknownProperties(array $properties, string $className): DataTransferObjectError
     {
         $propertyNames = implode('`, `', $properties);
 
         return new self("Public properties `{$propertyNames}` not found on {$className}");
     }
 
-    public static function invalidType(Property $property, $value): DataObjectError
+    public static function invalidType(Property $property, $value): DataTransferObjectError
     {
         if ($value === null) {
             $value = 'null';
@@ -32,7 +32,7 @@ class DataObjectError extends TypeError
         return new self("Invalid type: expected {$property->getFqn()} to be of type {$expectedTypes}, instead got value `{$value}`.");
     }
 
-    public static function uninitialized(Property $property): DataObjectError
+    public static function uninitialized(Property $property): DataTransferObjectError
     {
         return new self("Non-nullable property {$property->getFqn()} has not been initialized.");
     }