+++ tests/TestClasses/TestDataTransferObject.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Spatie\DataTransferObject\Tests\TestClasses;
+
+use Spatie\DataTransferObject\DataTransferObject;
+
+class TestDataTransferObject extends DataTransferObject
+{
+    /** @var int */
+    public $testProperty;
+}